refactor: agent core
