+++ app/admin/api.py
@@ -2,9 +2,6 @@ import importlib
 import logging
 from typing import Annotated, Optional, TypedDict
 
-from aiogram import Bot
-from aiogram.exceptions import TelegramConflictError, TelegramUnauthorizedError
-from aiogram.utils.token import TokenValidationError
 from fastapi import (
     APIRouter,
     Body,
@@ -25,14 +22,16 @@ from sqlalchemy.orm.exc import NoResultFound
 from yaml import safe_load
 
 from app.auth import verify_admin_jwt
-from intentkit.clients.cdp import get_cdp_client
 from intentkit.clients.twitter import unlink_twitter
-from intentkit.config.config import config
+from intentkit.core.agent import (
+    _process_agent_post_actions,
+    _process_telegram_config,
+    create_agent,
+    override_agent,
+)
 from intentkit.core.engine import clean_agent_memory
-from intentkit.core.skill import skill_store
 from intentkit.models.agent import (
     Agent,
-    AgentCreate,
     AgentResponse,
     AgentTable,
     AgentUpdate,
@@ -41,213 +40,11 @@ from intentkit.models.agent_data import AgentData, AgentDataTable
 from intentkit.models.db import get_db
 from intentkit.models.user import User
 from intentkit.skills import __all__ as skill_categories
-from intentkit.utils.slack_alert import send_slack_message
 
 admin_router_readonly = APIRouter()
 admin_router = APIRouter()
 
 logger = logging.getLogger(__name__)
-
-
-async def _process_agent_post_actions(
-    agent: Agent, is_new: bool = True, slack_message: str | None = None
-) -> AgentData:
-    """Process common actions after agent creation or update.
-
-    Args:
-        agent: The agent that was created or updated
-        is_new: Whether the agent is newly created
-        slack_message: Optional custom message for Slack notification
-
-    Returns:
-        AgentData: The processed agent data
-    """
-    if config.cdp_api_key_id and agent.wallet_provider == "cdp":
-        cdp_client = await get_cdp_client(agent.id, skill_store)
-        await cdp_client.get_wallet_provider()
-
-    # Get new agent data
-    # FIXME: refuse to change wallet provider
-    if agent.wallet_provider == "readonly":
-        agent_data = await AgentData.patch(
-            agent.id,
-            {
-                "evm_wallet_address": agent.readonly_wallet_address,
-            },
-        )
-    else:
-        agent_data = await AgentData.get(agent.id)
-
-    # Send Slack notification
-    slack_message = slack_message or ("Agent Created" if is_new else "Agent Updated")
-    try:
-        _send_agent_notification(agent, agent_data, slack_message)
-    except Exception as e:
-        logger.error("Failed to send Slack notification: %s", e)
-
-    return agent_data
-
-
-async def _process_telegram_config(
-    agent: AgentUpdate, existing_agent: Optional[Agent], agent_data: AgentData
-) -> AgentData:
-    """Process telegram configuration for an agent.
-
-    Args:
-        agent: The agent with telegram configuration
-        agent_data: The agent data to update
-
-    Returns:
-        AgentData: The updated agent data
-    """
-    changes = agent.model_dump(exclude_unset=True)
-    if not changes.get("telegram_entrypoint_enabled"):
-        return agent_data
-
-    if not changes.get("telegram_config") or not changes.get("telegram_config").get(
-        "token"
-    ):
-        return agent_data
-
-    tg_bot_token = changes.get("telegram_config").get("token")
-
-    if existing_agent and existing_agent.telegram_config.get("token") == tg_bot_token:
-        return agent_data
-
-    try:
-        bot = Bot(token=tg_bot_token)
-        bot_info = await bot.get_me()
-        agent_data.telegram_id = str(bot_info.id)
-        agent_data.telegram_username = bot_info.username
-        agent_data.telegram_name = bot_info.first_name
-        if bot_info.last_name:
-            agent_data.telegram_name = f"{bot_info.first_name} {bot_info.last_name}"
-        await agent_data.save()
-        try:
-            await bot.close()
-        except Exception:
-            pass
-        return agent_data
-    except (
-        TelegramUnauthorizedError,
-        TelegramConflictError,
-        TokenValidationError,
-    ) as req_err:
-        logger.error(
-            f"Unauthorized err getting telegram bot username with token {tg_bot_token}: {req_err}",
-        )
-        return agent_data
-    except Exception as e:
-        logger.error(
-            f"Error getting telegram bot username with token {tg_bot_token}: {e}",
-        )
-        return agent_data
-
-
-def _send_agent_notification(agent: Agent, agent_data: AgentData, message: str) -> None:
-    """Send a notification about agent creation or update.
-
-    Args:
-        agent: The agent that was created or updated
-        agent_data: The agent data to update
-        message: The notification message
-    """
-    # Format autonomous configurations - show only enabled ones with their id, name, and schedule
-    autonomous_formatted = ""
-    if agent.autonomous:
-        enabled_autonomous = [auto for auto in agent.autonomous if auto.enabled]
-        if enabled_autonomous:
-            autonomous_items = []
-            for auto in enabled_autonomous:
-                schedule = (
-                    f"cron: {auto.cron}" if auto.cron else f"minutes: {auto.minutes}"
-                )
-                autonomous_items.append(
-                    f"• {auto.id}: {auto.name or 'Unnamed'} ({schedule})"
-                )
-            autonomous_formatted = "\n".join(autonomous_items)
-        else:
-            autonomous_formatted = "No enabled autonomous configurations"
-    else:
-        autonomous_formatted = "None"
-
-    # Format skills - find categories with enabled: true and list skills in public/private states
-    skills_formatted = ""
-    if agent.skills:
-        enabled_categories = []
-        for category, skill_config in agent.skills.items():
-            if skill_config and skill_config.get("enabled") is True:
-                skills_list = []
-                states = skill_config.get("states", {})
-                public_skills = [
-                    skill for skill, state in states.items() if state == "public"
-                ]
-                private_skills = [
-                    skill for skill, state in states.items() if state == "private"
-                ]
-
-                if public_skills:
-                    skills_list.append(f"  Public: {', '.join(public_skills)}")
-                if private_skills:
-                    skills_list.append(f"  Private: {', '.join(private_skills)}")
-
-                if skills_list:
-                    enabled_categories.append(
-                        f"• {category}:\n{chr(10).join(skills_list)}"
-                    )
-
-        if enabled_categories:
-            skills_formatted = "\n".join(enabled_categories)
-        else:
-            skills_formatted = "No enabled skills"
-    else:
-        skills_formatted = "None"
-
-    send_slack_message(
-        message,
-        attachments=[
-            {
-                "color": "good",
-                "fields": [
-                    {"title": "ID", "short": True, "value": agent.id},
-                    {"title": "Name", "short": True, "value": agent.name},
-                    {"title": "Model", "short": True, "value": agent.model},
-                    {
-                        "title": "Network",
-                        "short": True,
-                        "value": agent.network_id or agent.cdp_network_id or "Default",
-                    },
-                    {
-                        "title": "X Username",
-                        "short": True,
-                        "value": agent_data.twitter_username,
-                    },
-                    {
-                        "title": "Telegram Enabled",
-                        "short": True,
-                        "value": str(agent.telegram_entrypoint_enabled),
-                    },
-                    {
-                        "title": "Telegram Username",
-                        "short": True,
-                        "value": agent_data.telegram_username,
-                    },
-                    {
-                        "title": "Wallet Address",
-                        "value": agent_data.evm_wallet_address,
-                    },
-                    {
-                        "title": "Autonomous",
-                        "value": autonomous_formatted,
-                    },
-                    {
-                        "title": "Skills",
-                        "value": skills_formatted,
-                    },
-                ],
-            }
-        ],
-    )
 
 
 @admin_router_readonly.post(
@@ -335,20 +132,15 @@ async def validate_agent_update(
 
 
 @admin_router.post(
-    "/agents/v2",
+    "/agents",
     tags=["Agent"],
+    status_code=201,
     operation_id="create_agent",
-    summary="Create Agent",
-    response_model=AgentResponse,
     responses={
-        200: {"model": AgentResponse, "description": "Agent already exists"},
-        201: {"model": AgentResponse, "description": "Agent created"},
-        400: {"description": "Other client errors except format error"},
-        422: {"description": "Invalid agent configuration"},
-        500: {"description": "Server error"},
+        201: {"description": "Agent created successfully"},
     },
 )
-async def create_agent(
+async def create_agent_endpoint(
     input: AgentUpdate = Body(AgentUpdate, description="Agent configuration"),
     subject: str = Depends(verify_admin_jwt),
 ) -> Response:
@@ -365,34 +157,14 @@ async def create_agent(
         - 400: Invalid agent ID format or agent ID already exists
         - 500: Database error
     """
-    agent = AgentCreate.model_validate(input)
-    if subject:
-        agent.owner = subject
+    agent_response = await create_agent(input, subject)
 
-    # Check for existing agent by upstream_id
-    existing = await agent.get_by_upstream_id()
-    if existing:
-        agent_data = await AgentData.get(existing.id)
-        agent_response = await AgentResponse.from_agent(existing, agent_data)
-        return Response(
-            status_code=200,
-            content=agent_response.model_dump_json(),
-            media_type="application/json",
-            headers={"ETag": agent_response.etag()},
-        )
-    # Create new agent
-    latest_agent = await agent.create()
-    # Process common post-creation actions
-    agent_data = await _process_agent_post_actions(latest_agent, True, "Agent Created")
-    agent_data = await _process_telegram_config(input, None, agent_data)
-    agent_response = await AgentResponse.from_agent(latest_agent, agent_data)
-
-    # Return Response with ETag header
+    # Return Response with ETag header and appropriate status code
     return Response(
-        status_code=201,
         content=agent_response.model_dump_json(),
         media_type="application/json",
         headers={"ETag": agent_response.etag()},
+        status_code=201,
     )
 
 
@@ -457,7 +229,7 @@ async def update_agent(
 @admin_router.put(
     "/agents/{agent_id}", tags=["Agent"], status_code=200, operation_id="override_agent"
 )
-async def override_agent(
+async def override_agent_endpoint(
     agent_id: str = Path(..., description="ID of the agent to update"),
     agent: AgentUpdate = Body(AgentUpdate, description="Agent update configuration"),
     subject: str = Depends(verify_admin_jwt),
@@ -482,27 +254,7 @@ async def override_agent(
         - 403: Permission denied (if owner mismatch)
         - 500: Database error
     """
-    if subject:
-        agent.owner = subject
-
-    if not agent.owner:
-        raise HTTPException(status_code=400, detail="Owner is required")
-
-    existing_agent = await Agent.get(agent_id)
-    if not existing_agent:
-        raise HTTPException(status_code=404, detail="Agent not found")
-
-    # Update agent
-    latest_agent = await agent.override(agent_id)
-
-    # Process common post-update actions
-    agent_data = await _process_agent_post_actions(
-        latest_agent, False, "Agent Overridden"
-    )
-
-    agent_data = await _process_telegram_config(agent, existing_agent, agent_data)
-
-    agent_response = await AgentResponse.from_agent(latest_agent, agent_data)
+    agent_response = await override_agent(agent_id, agent, subject)
 
     # Return Response with ETag header
     return Response(
